+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## 1.2.4
 - 设置 minSdkVersion = 21
 - remove setUseSystemBarTintLollipop()
+- Deprecated setStatusBarTextColor(),合并于setStatusBarTintColor方法中
 - date：